+++ packages/quill/test/unit/modules/toolbar.spec.ts
@@ -178,12 +178,10 @@ describe('Toolbar', () => {
       ) as HTMLButtonElement;
       quill.setSelection(7);
       expect(boldButton.classList.contains('ql-active')).toBe(true);
-      // @ts-expect-error -- TODO fix later
-      expect(boldButton.attributes['aria-pressed'].value).toBe('true');
+      expect(boldButton.getAttribute('aria-pressed')).toBe('true');
       quill.setSelection(2);
       expect(boldButton.classList.contains('ql-active')).toBe(false);
-      // @ts-expect-error -- TODO fix later
-      expect(boldButton.attributes['aria-pressed'].value).toBe('false');
+      expect(boldButton.getAttribute('aria-pressed')).toBe('false');
     });
 
     test('link', () => {
@@ -193,12 +191,10 @@ describe('Toolbar', () => {
       ) as HTMLButtonElement;
       quill.setSelection(12);
       expect(linkButton.classList.contains('ql-active')).toBe(true);
-      // @ts-expect-error -- TODO fix later
-      expect(linkButton.attributes['aria-pressed'].value).toBe('true');
+      expect(linkButton.getAttribute('aria-pressed')).toBe('true');
       quill.setSelection(2);
       expect(linkButton.classList.contains('ql-active')).toBe(false);
-      // @ts-expect-error -- TODO fix later
-      expect(linkButton.attributes['aria-pressed'].value).toBe('false');
+      expect(linkButton.getAttribute('aria-pressed')).toBe('false');
     });
 
     test('dropdown', () => {
@@ -227,24 +223,27 @@ describe('Toolbar', () => {
       quill.setSelection(17);
       expect(centerButton.classList.contains('ql-active')).toBe(true);
       expect(leftButton.classList.contains('ql-active')).toBe(false);
-      // @ts-expect-error -- TODO fix later
-      expect(centerButton.attributes['aria-pressed'].value).toBe('true');
-      // @ts-expect-error -- TODO fix later
-      expect(leftButton.attributes['aria-pressed'].value).toBe('false');
+      expect(centerButton.getAttribute('aria-pressed')).toBe('true');
+      expect(leftButton.getAttribute('aria-pressed')).toBe('false');
       quill.setSelection(2);
       expect(centerButton.classList.contains('ql-active')).toBe(false);
       expect(leftButton.classList.contains('ql-active')).toBe(true);
-      // @ts-expect-error -- TODO fix later
-      expect(centerButton.attributes['aria-pressed'].value).toBe('false');
-      // @ts-expect-error -- TODO fix later
-      expect(leftButton.attributes['aria-pressed'].value).toBe('true');
+      expect(centerButton.getAttribute('aria-pressed')).toBe('false');
+      expect(leftButton.getAttribute('aria-pressed')).toBe('true');
       quill.blur();
       expect(centerButton.classList.contains('ql-active')).toBe(false);
       expect(leftButton.classList.contains('ql-active')).toBe(false);
-      // @ts-expect-error -- TODO fix later
-      expect(centerButton.attributes['aria-pressed'].value).toBe('false');
-      // @ts-expect-error -- TODO fix later
-      expect(leftButton.attributes['aria-pressed'].value).toBe('false');
+      expect(centerButton.getAttribute('aria-pressed')).toBe('false');
+      expect(leftButton.getAttribute('aria-pressed')).toBe('false');
+    });
+
+    test('update on format', function () {
+      const { container, quill } = setup();
+      const boldButton = container?.parentNode?.querySelector('button.ql-bold');
+      quill.setSelection(1, 2);
+      expect(boldButton?.classList.contains('ql-active')).toBe(false);
+      quill.format('bold', true, 'user');
+      expect(boldButton?.classList.contains('ql-active')).toBe(true);
     });
   });
 });