+++ CHANGELOG.md
@@ -1,10 +1,28 @@
 # [Unreleased]
 
+# 2.0.0-rc.0
+
+- **Clipboard** Convert newlines between inline elements to a space.
+- **Clipboard** Avoid generating unsupported formats on paste.
+- **Clipboard** Improve support for pasting from Google Docs and Microsoft Word.
+- **Clipboard** Ignore whitespace between pasted empty paragraphs.
+- **Syntax** Support highlight.js v10 and v11.
+
+# 2.0.0-beta.2
+
+- Fix IME not working correctly in Safari.
+- **Clipboard** Support paste as plain text.
+- Fix `Quill.getText()` not respecting `length` parameter.
+- **History** Fix redo shortcut not working on Linux and Windows.
+
+# 2.0.0-beta.1
+
 - Fix syntax label from "Javascript" to "JavaScript".
 - Fix typing errors for emitter.
 - Inline SVG images for easier bundler setup.
+- Improve typing for Registry.
 
-# v2.0.0-beta.0
+# 2.0.0-beta.0
 
 In the upcoming 2.0 release, Quill has been significantly modernized. Leveraging the latest browser-supported APIs, Quill now delivers a more efficient and reliable editing experience.